Move tests using curator testing server to more general package.
