+++ test/src/main/java/org/apache/accumulo/test/zookeeper/ZooKeeperTestingServer.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.master.metrics.fate;
+package org.apache.accumulo.test.zookeeper;
 
 import java.io.IOException;
 import java.nio.file.Files;
